Implemented WP coding standards

+++ includes/Rest/ActionsController.php
@@ -104,6 +104,11 @@ class ActionsController extends BaseController {
 			return $this->bad_request( 'title (or label) is required.' );
 		}
 
+		$cap = $this->assert_can_create_post_type( 'wpail_action' );
+		if ( is_wp_error( $cap ) ) {
+			return $cap;
+		}
+
 		$post_id = wp_insert_post( [
 			'post_type'   => 'wpail_action',
 			'post_title'  => $title,
@@ -132,6 +137,11 @@ class ActionsController extends BaseController {
 		}
 
 		$post_id  = $action->id;
+		$cap      = $this->assert_can_edit_post( $post_id );
+		if ( is_wp_error( $cap ) ) {
+			return $cap;
+		}
+
 		$params   = (array) ( $request->get_json_params() ?? [] );
 		$old_meta = RelationshipHelper::get_meta( $post_id );
 
@@ -156,6 +166,11 @@ class ActionsController extends BaseController {
 		}
 
 		$post_id  = $action->id;
+		$cap      = $this->assert_can_delete_post( $post_id );
+		if ( is_wp_error( $cap ) ) {
+			return $cap;
+		}
+
 		$old_meta = RelationshipHelper::get_meta( $post_id );
 		RelationshipSync::sync( $post_id, 'wpail_action', $old_meta, [] );
 		wp_delete_post( $post_id, true );
