+++ includes/Rest/FaqsController.php
@@ -121,6 +121,11 @@ class FaqsController extends BaseController {
 			return $this->bad_request( 'short_answer is required.' );
 		}
 
+		$cap = $this->assert_can_create_post_type( 'wpail_faq' );
+		if ( is_wp_error( $cap ) ) {
+			return $cap;
+		}
+
 		$post_id = wp_insert_post( [
 			'post_type'   => 'wpail_faq',
 			'post_title'  => $question,
@@ -149,6 +154,11 @@ class FaqsController extends BaseController {
 		}
 
 		$post_id  = $faq->id;
+		$cap      = $this->assert_can_edit_post( $post_id );
+		if ( is_wp_error( $cap ) ) {
+			return $cap;
+		}
+
 		$params   = (array) ( $request->get_json_params() ?? [] );
 		$old_meta = RelationshipHelper::get_meta( $post_id );
 
@@ -175,6 +185,11 @@ class FaqsController extends BaseController {
 		}
 
 		$post_id  = $faq->id;
+		$cap      = $this->assert_can_delete_post( $post_id );
+		if ( is_wp_error( $cap ) ) {
+			return $cap;
+		}
+
 		$old_meta = RelationshipHelper::get_meta( $post_id );
 		RelationshipSync::sync( $post_id, 'wpail_faq', $old_meta, [] );
 		wp_delete_post( $post_id, true );
