+++ includes/Rest/ServicesController.php
@@ -100,6 +100,11 @@ class ServicesController extends BaseController {
 			return $this->bad_request( 'title is required.' );
 		}
 
+		$cap = $this->assert_can_create_post_type( 'wpail_service' );
+		if ( is_wp_error( $cap ) ) {
+			return $cap;
+		}
+
 		$post_id = wp_insert_post( [
 			'post_type'   => 'wpail_service',
 			'post_title'  => $title,
@@ -128,6 +133,11 @@ class ServicesController extends BaseController {
 		}
 
 		$post_id  = $service->id;
+		$cap      = $this->assert_can_edit_post( $post_id );
+		if ( is_wp_error( $cap ) ) {
+			return $cap;
+		}
+
 		$params   = (array) ( $request->get_json_params() ?? [] );
 		$old_meta = RelationshipHelper::get_meta( $post_id );
 
@@ -152,6 +162,11 @@ class ServicesController extends BaseController {
 		}
 
 		$post_id  = $service->id;
+		$cap      = $this->assert_can_delete_post( $post_id );
+		if ( is_wp_error( $cap ) ) {
+			return $cap;
+		}
+
 		$old_meta = RelationshipHelper::get_meta( $post_id );
 		RelationshipSync::sync( $post_id, 'wpail_service', $old_meta, [] );
 		wp_delete_post( $post_id, true );
